+++ datastore/tests/test_permission_triggers.py
@@ -2,121 +2,8 @@
 Test that object inserts and removals trigger updates to
 permission_object_mapping
 """
-from collections import OrderedDict
-
-
 import pymysql
 import pytest
-
-
-from conftest import newuuid
-
-
-@pytest.fixture(scope='session')
-def anint():
-    i = 0
-    yield i
-    i += 1
-
-
-@pytest.fixture(scope='function')
-def new_organization(cursor, anint):
-    def fnc(i=anint):
-        out = OrderedDict(id=newuuid(), name=f'org{i}')
-        cursor.execute('INSERT INTO organizations (id, name) VALUES (%s, %s)',
-                       list(out.values()))
-        return out
-    return fnc
-
-
-@pytest.fixture()
-def new_user(cursor, new_organization, anint):
-    def fcn(org=None, i=anint):
-        if org is None:
-            org = new_organization(i=i)
-        out = OrderedDict(id=newuuid(), auth0_id=f'authid{i}',
-                          organization_id=org['id'])
-        cursor.execute(
-            'INSERT INTO users (id, auth0_id, organization_id) VALUES '
-            '(%s, %s, %s)', list(out.values()))
-        return out
-    return fcn
-
-
-@pytest.fixture()
-def new_role(cursor, new_organization, anint):
-    def fcn(org=None, i=anint):
-        if org is None:
-            org = new_organization(i=i)
-        out = OrderedDict(id=newuuid(), name=f'role{i}',
-                          description='therole',
-                          organization_id=org['id'])
-        cursor.execute(
-            'INSERT INTO roles (id, name, description, organization_id) '
-            'VALUES (%s, %s, %s, %s)', list(out.values()))
-        return out
-    return fcn
-
-
-@pytest.fixture()
-def new_permission(cursor, new_organization, anint):
-    def fcn(action, object_type, applies_to_all, org=None, i=anint):
-        if org is None:
-            org = new_organization(i=i)
-        out = OrderedDict(id=newuuid(), description='perm',
-                          organization_id=org['id'],
-                          action=action, object_type=object_type,
-                          applies_to_all=applies_to_all)
-        cursor.execute(
-            'INSERT INTO permissions (id, description, organization_id, action'
-            ', object_type, applies_to_all) VALUES (%s, %s, %s, %s, %s, %s)',
-            list(out.values()))
-        return out
-    return fcn
-
-
-@pytest.fixture()
-def new_site(cursor, new_organization, anint):
-    def fcn(org=None, i=anint):
-        if org is None:
-            org = new_organization(i=i)
-        out = OrderedDict(id=newuuid(), organization_id=org['id'],
-                          name=f'site{i}')
-        cursor.execute(
-            'INSERT INTO sites (id, organization_id, name) VALUES '
-            '(%s, %s, %s)', list(out.values()))
-        return out
-    return fcn
-
-
-@pytest.fixture()
-def new_forecast(cursor, new_site, anint):
-    def fcn(site=None, org=None, i=anint):
-        if site is None:
-            site = new_site(org)
-        out = OrderedDict(
-            id=newuuid(), organization_id=site['organization_id'],
-            site_id=site['id'], name=f'forecast{i}')
-        cursor.execute(
-            'INSERT INTO forecasts (id, organization_id, site_id, name) VALUES '
-            '(%s, %s, %s, %s)', list(out.values()))
-        return out
-    return fcn
-
-
-@pytest.fixture()
-def new_observation(cursor, new_site, anint):
-    def fcn(site=None, org=None, i=anint):
-        if site is None:
-            site = new_site(org)
-        out = OrderedDict(
-            id=newuuid(), organization_id=site['organization_id'],
-            site_id=site['id'], name=f'observation{i}')
-        cursor.execute(
-            'INSERT INTO observations (id, organization_id, site_id, name) '
-            'VALUES (%s, %s, %s, %s)', list(out.values()))
-        return out
-    return fcn
 
 
 def test_permissions_not_updatable(cursor, new_permission):
@@ -143,22 +30,6 @@ def test_object_mapping_new_permission(cursor, new_permission,
         'permission_id = %s', perm1['id'])
     # 2 includes first permission and this second one with applies_to_all
     assert cursor.fetchone()[0] == 2
-
-
-@pytest.fixture(params=['sites', 'users', 'roles', 'forecasts',
-                        'observations'])
-def getfcn(request, new_site, new_user, new_role, new_forecast,
-           new_observation):
-    if request.param == 'sites':
-        return new_site, 'sites'
-    elif request.param == 'users':
-        return new_user, 'users'
-    elif request.param == 'roles':
-        return new_role, 'roles'
-    elif request.param == 'forecasts':
-        return new_forecast, 'forecasts'
-    elif request.param == 'observations':
-        return new_observation, 'observations'
 
 
 @pytest.mark.parametrize('action', ['read', 'update', 'delete'])
